% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/afmContactPoint.R
\name{afmContactPoint}
\alias{afmContactPoint}
\title{Find the contact point in for the Force-Distance curve
following the method described in Microscopy Research and Technique 2013}
\usage{
afmContactPoint(afmdata, width = 1, mul1, mul2, Delta = TRUE)
}
\arguments{
\item{afmdata}{A Force-Distance curve with the afmdata structure. It should be a list with at least the 'data' field with a data frame of at least 4 columns.}

\item{width}{Width of the window for the local regression (in vector position units)}

\item{mul1}{First multiplier for the first alarm threshold}

\item{mul2}{Second multiplier for the second alarm threshold}

\item{Delta}{Logical. If TRUE, then the statistic for determining the contact point is the differences between two consecutive values of the slope of the local regression line. If FALSE then the slope itself is used.}
}
\value{
A list of: CP The contact point value. iCP The position in the array for the contact point value. delta The delta signal. noise The noise of the delta signal
}
\description{
Find the contact point in for the Force-Distance curve
following the method described in Microscopy Research and Technique 2013
}

