% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afmdata.R
\name{afmdata}
\alias{afmdata}
\title{AFM data}
\usage{
afmdata(data, dstr = "Z", Fstr = "Force", Segstr = "Segment",
  params = list(SpringConstant = numeric()))
}
\arguments{
\item{\code{data}:}{A data frame consisting in 3 or 4 columns. A minimum of 
"Z" (or "distance"), "Force" and "Segment". Optionally a fourth column with 
"Time" could be added.}

\item{\code{dstr}:}{Character string with the posible names for the distance
variable.}

\item{\code{Fstr}:}{Character string with the posible names for the force 
variable.}

\item{\code{Segstr}:}{Character string with the posible names for the Segment
variable.}

\item{\code{tstr}:}{Character string with the posible names for the time
variable.}
}
\value{
An object of class \code{afmdata}
}
\description{
This function creates an \code{afmdata} structure, which is as list with at
 least one field called \code{data} which is a data frame with a valid AFM 
 data, that is, at least 3 variables called "Z", "Force", and "Segment".
}

