% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/afmExpDecay.R
\name{afmExpDecay}
\alias{afmExpDecay}
\title{Exponential decay fit}
\usage{
afmExpDecay(afmdata, nexp = 2, tmax = NULL, type = c("CH", "CF"),
  plt = TRUE, ...)
}
\arguments{
\item{\code{afmdata}:}{An object of \code{afmdata} class with a
\bold{Contact} segment and a \bold{Time} column in the \code{data}
dataframe.}

\item{\code{nexp}:}{Number of expontials in the Prony series to be fitted.
Currently only one or two exponentials are supported. Default is 2.}

\item{\code{tmax}:}{Maximum time considered in the relaxation curve. It
defaults to \code{Inf}, meaning that the whole contact segment is
considered.}

\item{\code{type}:}{Type of the experiment. Can be either "CH" (Constant
Height) for a force-relaxation experiment or "CF" (Constant Force) for a
creep experiment. Default is \code{type = "CH"}.}

\item{\code{plt:}}{Logical. If TRUE (default) then a plot of the Contact
segment with the overlay of the fit is shown.}

\item{\code{...}:}{Options passed to the \code{nlsM()} function from the
  \code{minpack.lm} package. At least should contain the starting values
  (\code{start = list(...)}) for the Levenberg-Mardquart nonlinear least
  square method.}
}
\value{
A list of:

  \code{expdecayModel}: A nls object returned from  \code{nlsM()} function.

  \code{expdecayFit}: The values predicted by the fit, returned from the
  \code{predict()} function.
}
\description{
Fits a viscoelastic exponential decay in a Force-Relaxation or
Creep experiments.
}

