% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afmexperiment.R
\name{afmexperiment}
\alias{afmexperiment}
\title{AFM experiment}
\usage{
afmexperiment(data, ID=NULL)
}
\arguments{
\item{data}{A variable of  \code{afmdata} class, or a list of elements of
\code{afmdata} class.}

\item{ID}{Character string with the identifier of the \code{data} variable or a string
array in case \code{data} is a list of \code{afmdata} variables.}
}
\value{
An object of class \code{afmexp}.
}
\description{
This function creates an \code{afmexperiment} structure, which is as list 
(or an array) of elements of \code{afmdata} class.
}
\examples{
dataFolder <- paste(path.package("afmToolkit"), "afmexperiment",sep = "/")
dataFiles <- list.files(dataFolder, pattern = "force", full.names = FALSE)
data <- lapply(dataFiles, afmReadJPK, path = dataFolder)
names(data) <- dataFiles
data <- afmexperiment(data)
plot(data[[1]])
}
\seealso{
\code{\link{afmdata}}
}
