% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afmExtract.R
\name{afmExtract}
\alias{afmExtract}
\title{Extract computed parameters from an afmexperiment}
\usage{
afmExtract(afmexperiment, params = list("YM", "AE", "ED"))
}
\arguments{
\item{afmexperiment}{Data of afmexperiment class.}

\item{params}{List of parameters to extract from the data.}

\item{opt.param}{Optional parameter or factor in the params field of the afmdata list to add to the data extraction.}
}
\value{
A data frame with the name of the curve and the corresponding values of the parameters extacted.
}
\description{
Extracts some parameters from an afmexperiment for an easy further analysis.
}
\examples{

TODO
ADD EXAMPLES

}

