% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/afmIndentation.R
\name{afmIndentation}
\alias{afmIndentation}
\title{afmIndentation}
\usage{
afmIndentation(afmdata)
}
\arguments{
\item{\code{afmdata}:}{An \code{afmdata} object. It should be a valid afmdata
 object upon which the Contact Point, the baseline correction and the Zero
 Force Point must have been calculated first (using functions
 \code{afmContactPoint()}, \code{afmBaselineCorrection()}) and
  \code{afmZeroPointSlope()}}
}
\value{
Returns a list with one field:

\code{afmdata}: An afmdata class in which a Deformation vector is added
in the \code{data} field
}
\description{
This function computes the deformation of the sample from the
calibrated Force-Distance curve, by substracting Z to the Zero Force Point
calculated with afmZeroPointSlope function.
}

