% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afmReadJPK.R
\name{afmReadJPK}
\alias{afmReadJPK}
\title{Read Nanowizard JPK ascii file}
\usage{
afmReadJPK(filename, path = ".", FColStr = "Vertical", ZColStr = "smoothed)", tColStr = "Series")
}
\arguments{
\item{filename}{String with the name of the jpk file}
}
\value{
A list containing a field 'data' which is a data frame
}
\description{
Read an ascii JPK file.

Reads an ascii JPK file with one to three headers.
}
\examples{
data <- afmReadJPK("force-save-JPK-3h.txt",path = path.package("afmToolkit"))
str(data)
}

