% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afmReadJPKMultiFolder.R
\name{afmReadJPKMultiFolder}
\alias{afmReadJPKMultiFolder}
\title{Read all Nanowizard Multi-indentation JPK ascii files in a folder}
\usage{
afmReadJPKMultiFolder(folder, pattern = ".txt", ...)
}
\arguments{
\item{folder}{Name of the  folder containing the jpk files.}

\item{pattern}{Pattern that will identify the jok files (".txt" by default).}

\item{...}{Other parameters passed to afmReadJPKMultiIndent function.}
}
\value{
An \code{afmmultiexp} class data structure with all F-d curves.
}
\description{
Read all multi-indentation JPK ascii files in a given folder. It searches for all files containing a 
given patter (".txt" by default) and uses the \code{afmReadJPJ} function.
}
\examples{
folder <- paste(path.package("afmToolkit"), "multiIndentExperiment",sep = "/")
data <- afmReadJPKMultiFolder(folder = folder)

}
