% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afmStitch.R
\name{afmStitch}
\alias{afmStitch}
\title{Stitch the files of a Bruker experiment}
\usage{
afmStitch(expname)
}
\arguments{
\item{expname}{Veeco experiment namea character string with the common part of the 4 files.}
}
\value{
An \code{afmdata} class variable with the three segments of the Veeco file experiment
}
\description{
Creates an afmdata structure from the four files of a Bruker experiment containing three segments:
approach, hold and retract. It also requires the header file to extract the spring constant.
}
