% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/afmYoungModulus.R
\name{afmYoungModulus}
\alias{afmYoungModulus}
\title{afmYoungModulus}
\usage{
afmYoungModulus(afmdata, model = "Hertz", geometry = "pyramid",
  silent = TRUE, params)
}
\arguments{
\item{\code{afmdata}:}{An \code{afmdata} object. It should be a valid afmdata
 object upon which the Contact Point, the baseline correction and the Zero
 Force Point must have been calculated first (using functions
 \code{afmContactPoint()}, \code{afmBaselineCorrection()}) and
  \code{afmZeroPointSlope()}}
}
\value{
Returns a list with one field:

\code{afmdata}: An afmdata class in which a Deformation vector is added
in the \code{data} field
}
\description{
This function computes the Young's Modulus of the sample
from the approach curve using different contact models and for different
tip geometries.
}

