% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afmZeroPointSlope.R
\name{afmZeroPointSlope}
\alias{afmZeroPointSlope}
\title{Zero Force Point and Slope}
\usage{
afmZeroPointSlope(afmdata, segment = c("approach", "retract"))
}
\arguments{
\item{\code{afmdata}:}{An \code{afmdata} object. It should be a valid afmdata object upon
which the Contact Point and the baseline correction must have been calculated first
(using functions \code{afmContactPoint()} and \code{afmBaselineCorrection()})}

\item{\code{segment}:}{The segment on which everything is calculated.}
}
\value{
Returns a list with two fields:

\code{Z0Point}: Point of zero force.
\code{Slope}: Slope of the best fit line in the contact part of the Force-Distance curve.
}
\description{
This function finds the point of zero force (real contact point)
and the slope of the contact part of the Force-Distance curve.
}

