% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is.afmdata.R
\name{is.afmdata}
\alias{is.afmdata}
\title{Afmdata check.}
\usage{
is.afmdata(x)
}
\arguments{
\item{\code{x}:}{Any \bold{R} object.}
}
\value{
Returns \code{TRUE} if its argument is an afmdata (that is, has "afmdata"
amongst its classes) and \code{FALSE} otherwise.
}
\description{
Checks wether an R object is an afmdata or not.
}

