% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.afmexperiment.R
\name{is.afmexperiment}
\alias{is.afmexperiment}
\title{Afmexperiment check.}
\usage{
is.afmexperiment(x)
}
\arguments{
\item{x}{Any \bold{R} object.}
}
\value{
Returns \code{TRUE} if its argument is an afmdata (that is, has "afmexperiment"
  amongst its classes) and \code{FALSE} otherwise.
}
\description{
Checks wether an R object is an afmexperiment or not.
}
