% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.afmmulti.R
\name{is.afmmulti}
\alias{is.afmmulti}
\title{Afmdata multiindentation check.}
\usage{
is.afmmulti(x)
}
\arguments{
\item{x}{Any \bold{R} object.}
}
\value{
Returns \code{TRUE} if its argument is an afmmulti (that is, has "afmmulti" 
  amongst its classes) and \code{FALSE} otherwise.
}
\description{
Checks whether an R object is of class afmmulti or not.
}
