% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.afmdata.R
\name{plot.afmdata}
\alias{plot.afmdata}
\title{Plot an afmdata object}
\usage{
\method{plot}{afmdata}(afmdata, vs = "Z", segment = "all", ...)
}
\arguments{
\item{\code{afmdata}:}{An object of \code{afmdata} class.}

\item{\code{vs}:}{The variable for the x-axis. May take the values "Time" or 
"Z". It defaults to "Z", plotting thus a Force-Distance curve. If \code{vs}
is set to "Time", then it plots  a Force-Time curve.}

\item{\code{segment}:}{The segment of the curve to be plotted. If
\code{segment = "all"} then all segments of the curve are plotted. Possible
values are: \code{"approach"}, \code{"contact"}, \code{"retract"} and 
\code{"all"}.}
}
\description{
Plots an afmdata object.
}

