% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afmExtract.R
\name{afmExtract}
\alias{afmExtract}
\title{Extract computed parameters from an afmexperiment}
\usage{
afmExtract(afmexperiment, params = list("YM", "AE", "ED"))
}
\arguments{
\item{afmexperiment}{Data of afmexperiment class.}

\item{params}{List of parameters to extract from the data.}
}
\value{
A data frame with the name of the curve and the corresponding values of the parameters extacted.
}
\description{
Extracts some parameters from an afmexperiment for an easy further analysis.
}
\examples{
data <- afmReadJPK("force-save-JPK-3h.txt",path = path.package("afmToolkit"))
str(data)
}

