% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/afmReadJPK.R
\name{afmReadJPK}
\alias{afmReadJPK}
\title{Read an ascii JPK file.}
\usage{
afmReadJPK(filename, ZColStr = "Vertical", FColStr = "smoothed)",
  tColStr = "Series")
}
\arguments{
\item{filename}{String with the name of the jpk file}
}
\value{
A list containing a field 'data' which is a data frame
}
\description{
Reads an ascii JPK file with one to three headers.
}
\examples{
afmReadJPK('testJPKfile.txt')
}

