% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afmReadVeeco.R
\name{afmReadVeeco}
\alias{afmReadVeeco}
\title{Read Bruke Nanoscope Veeco ascii file}
\usage{
afmReadVeeco(filename, path = "")
}
\arguments{
\item{filename}{String with the name of the jpk file.}

\item{path}{Path to the folder where the file is.}
}
\value{
A list containing a field 'data' which is a data frame
}
\description{
Read an ascii Veeco file.

Reads an ascii Veeco file with one or two segments.
}
\examples{
data <- afmReadVeeco("veeco_file.txt.gz",path = path.package("afmToolkit"))
str(data)
}

