% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.afmdata.R
\name{append.afmdata}
\alias{append.afmdata}
\title{Append to an \code{afmdata} list.}
\usage{
append.afmdata(afmdata, x, name = NULL)
}
\arguments{
\item{afmdata}{The afmdata to which the new list is going to be joined.}

\item{x}{A list to be appended.}

\item{name}{The name of new field of the resulting afmdata object. If none is given, it
is the same as \code{x}.}
}
\value{
The new list of class \code{afmdata}
}
\description{
This function appends a list to an existing afmdata structure. It is used
internally by several afm* functions when attaching the results to the input afmdata
variable. This function should not be used directly unless by experienced users.
}
